pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mengrowth"
version = "0.1.0"
description = "MenGrowth: A deep learning framework for meningioma growth prediction from multimodal magnetic resonance imaging"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = [
    "magnetic-resonance-imaging",
    "deep-learning",
    "tumour-growth-prediction",
    "meningioma"
]
authors = [
    {name = "Mario Pascual-González", email = "mpascual@uma.es"},
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "PyYAML",
    "SimpleITK",
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "nibabel>=5.0.0",
    "nipype>=1.8.6",
    "pynrrd",
    "scipy",
    "scikit-image",
    "openpyxl"
]

[project.optional-dependencies]
preprocessing = [
    "brats==0.1.6",
    "brats[preprocessing]",
    "intensity-normalization @ git+https://github.com/jcreinhold/intensity-normalization.git@6d2e486",
    "cupy-cuda12x",
]

[project.urls]
Homepage = "https://github.com/MarioPasc/MenGrowth"
Documentation = "https://github.com/MarioPasc/MenGrowth#readme"
Repository = "https://github.com/MarioPasc/MenGrowth"
"Bug Tracker" = "https://github.com/MarioPasc/MenGrowth/issues"
Changelog = "https://github.com/MarioPasc/MenGrowth/releases"

[project.scripts]
mengrowth-reorganize = "mengrowth.cli.reorganize_raw_data:main"
mengrowth-filter = "mengrowth.cli.filter_raw_data:main"
mengrowth-analyze-dataset = "mengrowth.cli.analyze_dataset:main"
mengrowth-visualize-analysis = "mengrowth.cli.visualize_analysis:main"
mengrowth-curate-dataset = "mengrowth.cli.curate_dataset:main"
mengrowth-preprocess = "mengrowth.cli.preprocess:main"
mengrowth-segment = "mengrowth.cli.segment:main"

[tool.setuptools.packages.find]
include = ["mengrowth*"]

[tool.setuptools.package-data]
mengrowth = ["py.typed"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "yaml",
    "pandas",
    "SimpleITK.*",
    "matplotlib.*",
    "seaborn.*",
    "numpy.*",
    "nibabel.*",
    "nrrd.*",
]
ignore_missing_imports = true